requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
python-docx==0.8.11
fpdf2==2.7.4
pytesseract==0.3.10
pydub==0.25.1
vosk==0.3.45
google-generativeai==0.3.0
python-dotenv==1.0.0
Pillow==10.0.0
flask-cors==3.0.10
pdfminer.six==20221105
SpeechRecognition==3.10.0 